Added extra check-point for data directory in the StelFileMgr class

+++ src/core/StelFileMgr.cpp
@@ -103,7 +103,7 @@ void StelFileMgr::init()
 			ResourcesDir.cd(QString("Resources"));
 		}
 		QFileInfo installLocation(ResourcesDir.absolutePath());
-		QFileInfo checkFile(installLocation.filePath() + QString("/") + QString(CHECK_FILE));
+		QFileInfo checkFile(installLocation.filePath() + QDir::separator() + QString(CHECK_FILE));
 	#elif defined(Q_OS_WIN)		
 		QFileInfo installLocation(QCoreApplication::applicationDirPath());
 		QFileInfo checkFile(installLocation.filePath() + QDir::separator() + QString(CHECK_FILE));
@@ -148,10 +148,27 @@ void StelFileMgr::init()
 						 << QDir::toNativeSeparators(relativePath)
 						 << " (we checked for "
 						 << QDir::toNativeSeparators(checkFile.filePath()) << ").";
-				#ifndef UNIT_TEST
-				// NOTE: Hook for buildbots (using within testEphemeris)
-				qFatal("Couldn't find install directory location.");
-				#endif
+
+				qWarning() << "Maybe this is development environment? Let's check source directory path...";
+
+				QString sourceDirPath = STELLARIUM_SOURCE_DIR; // The variable is defined in CMakeLists.txt file
+				checkFile = QFileInfo(sourceDirPath + QDir::separator() + CHECK_FILE);
+				if (checkFile.exists())
+				{
+					installDir = sourceDirPath;
+				}
+				else
+				{
+					qWarning() << "WARNING StelFileMgr::StelFileMgr: could not find install location:"
+							 << QDir::toNativeSeparators(sourceDirPath)
+							 << " (we checked for "
+							 << QDir::toNativeSeparators(checkFile.filePath()) << ").";
+
+					#ifndef UNIT_TEST
+					// NOTE: Hook for buildbots (using within testEphemeris)
+					qFatal("Couldn't find install directory location.");
+					#endif
+				}
 			}
 		}
 	}
